Add a write of "1" to commandReady at the end of every tpm cmd write.
According to spec, two writes of "1" to commandReady bit are required. The first one should be immediately after receiving the response to transit TPM into idle state and to allow TPMs that perform background processing to proceed; the second one is transit idle state to ready state before sending the second command.

Signed-off-by: David Dorsey <trogdorsey@gmail.com>
Reviewed-by: Gang Wei <gang.wei@intel.com>
ACKed-by: Joseph Cihula <joseph.cihula@intel.com>