Fix boot of relocatable kernel
A reloctable kernel will uncompress into the memory just above the protected mode code, which may overwrite tboot.  When a relocatable kernel is being used, tboot will put the protected mode code after the tboot image (incl. the VT-d rounding) so that when it uncompresses it does not interfere with tboot.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>