Walkaround for pvops to read public config registers.
On Xen with pvops dom0, the public config registers can't be read because the address space is not mapped in pvops. However, they can be read on Xen with the previous 2.6.18 dom0.

Since the public config registers should be accessed always. This patch is a walkaround based on c/s 187.

Signed-off-by: Shane Wang <shane.wang@intel.com>