Add support for choosing correct SINIT ACM from multiple loaded modules
tboot will now go through all of the modules that are loaded by the bootloader and determine which ones are SINIT ACMs.  Of these, it will choose the one that matches the chipset (incl. debug/production fusing) and use that one.  All SINIT ACMs will be removed from the module list before executing the kernel/VMM.

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>