Fixed bug with PCONF elements not displaying correctly
Due to the definition of pcr_infos[] in lcp_pconf_element_t being of type uin8_t[], lcp_crtpolelt would not display multiple PCONF elements correctly.  To enable users of lcp2.h that use the lcp_pconf_element_t to get the correct type for pcr_infos[] (tpm_pcr_info_short_t), the type in lcp2.h was replaced with a macro that defaults to uint8_t (for most users, that don't care) and is re-defined to tpm_pcr_info_short_t in pconf_elt.c.

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>