Add support for TXT heap extended data elements and BiosData version 4 Add support for 'j' precision modifier to vcsnprintf()
The latest TXT MLE Developers Guide defines extended data elements that will be used for all future extensions to the TXT heap data structures.  The first such elements are used in the BiosData struct versions >= 4.  These elements are only informative, so tboot only validates and displays them.

The 'j' precision modifier is equivalent to 'L' or 'll' and is recognized by gcc as applyign to types of uint64_t.  This allows for code to be shared between tboot and Linux utilities (e.g. acminfo, txt-stat).

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>