Fix for GRUB2 loading elf image such as Xen.
Tboot assumes all modules are loaded in the memory above tboot memory. It is true for GRUB1, but false for GRUB2.
So for GRUB2 case, just move all modules & mbi in memory below tboot to the end of last module.

It just apply to elf image case such as Xen, and not apply to Linux. For Linux there is a separate patch to fix potential issue with GRUB2.

Signed-off-by: Gang Wei <gang.wei@intel.com>