Fix cmdline size in tb_polgen
There is a problem with tb_polgen. When it is parsing the kernel cmdline, it uses the first 256 characters of the cmdline, and discards the rest. Tboot itself, uses the first 1024 characters of the command line received from grub2. This means that the hash generated by tb_polgen will NOT match the hash generated by tboot/policy.c and verification will fail.

This patch fix it, and add a check in tb_polgen for the cmdline length being larger than the buffer so, and put out an error message to that effect so that the condition will be flagged.

Signed-off-by: Charles Fisher <Charles.Fisher@gdc4s.com>