Loading multiboot(2) images
We have been experimenting with measured launch of environments other than Linux and Xen, such
as the RTS hypervisor and some real-time operating systems.

We have seen a recurring issue in some cases with the way TBoot launches multiboot/multiboot2
environments. Namely, when loading the ELF image, TBoot overwrites the kernel ELF image and modules.

Developers at RTS have made changes to "tboot/common/loader.c" and "tboot/common/elf.c" that
addresses this issue:
When the follow-on kernel is a multiboot or multiboot2 kernel, move the modules to the top of
memory as reported in the E820 map. Then, after loading the kernel ELF binary into its final
location in memory, move the modules to a location just above the kernel.

Signed-off-by: Raphael Baum  <r.baum@real-time-systems.com>
Signed-off-by: Safayet Ahmed <Safayet.Ahmed@ge.com>
Signed-off-by: Ning Sun <ning.sun@intel.com>