If Vt-D is disabled in the BIOS, the DMAR table is not present.
We can look for this and skip trying to perform a measured launch.
This behavior of a missing DMAR table may not be true for all platforms,
but in any case if the DMAR table is not present the kernel will not be
able to detect/reprogram the IOMMU along the tboot_force_iommu path.
So either way this change should be helpful.

Testing:
Disable Vt-D, verify that measured launch is skipped.
Enable Vt-D, verify that measured launch is performed.


Signed-off-by: Sahil Rihan <srihan@fb.com>
Reviewed-by: Ning Sun <ning.sun@intel.com>