Make policy element stm_elt use unique type name
The stm_elt policy element plugin registers as type "mle" which
conflicts with the mle_elt policy plugin that also registers as type
"mle".

The plugin architecture uses C constructors to register plugins within
the framework.  When multiple compilation units use constructors, the
order in which the constructors are called is somewhat arbitrary.  It
depends on the compiler and the dynamic loader.

This patch registers the stm_elt as type "stm", which looks to have
been the original intent.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>
Reviewed-by: Ning Sun <ning.sun@intel.com>